Linux sdl2

+++ Editor/PaintToolWindow.cpp
@@ -4,6 +4,7 @@
 #include "ShaderInterop_Paint.h"
 
 #include <sstream>
+#include <cmath>
 
 using namespace wiECS;
 using namespace wiScene;
@@ -469,7 +470,7 @@ void PaintToolWindow::Update(float dt)
 				{
 					RecordHistory(true);
 					rebuild = true;
-					const float affection = amount * std::powf(1 - (dist / radius), falloff);
+					const float affection = amount * std::pow(1.0f - (dist / radius), falloff);
 
 					switch (mode)
 					{
@@ -596,7 +597,7 @@ void PaintToolWindow::Update(float dt)
 				if (z >= 0 && z <= 1 && dist <= radius)
 				{
 					averageNormal += N;
-					const float affection = amount * std::powf(1 - (dist / radius), falloff);
+					const float affection = amount * std::pow(1.0f - (dist / radius), falloff);
 					paintindices.push_back({ j, affection });
 				}
 			}
@@ -804,7 +805,7 @@ void PaintToolWindow::Update(float dt)
 					case MODE_HAIRPARTICLE_LENGTH:
 						if (hair->vertex_lengths[j] > 0) // don't change distribution
 						{
-							const float affection = amount * std::powf(1 - (dist / radius), falloff);
+							const float affection = amount * std::pow(1.0f - (dist / radius), falloff);
 							hair->vertex_lengths[j] = wiMath::Lerp(hair->vertex_lengths[j], color_float.w, affection);
 							// don't let it "remove" the vertex by keeping its length above zero:
 							//	(because if removed, distribution also changes which might be distracting)
